microservices/user-relations-service/consumer/requirements.txt
pika
python-dotenv
neo4j